Accommodate for Adhell3 removing * prefixes
Blocking was far too restrictive with the wildcard prefix.

A fair amount of head scratching was required to think of how to keep the host file size down.

Adjusted the reverse host lookup so that we count how many instances there are. After x instances, we output a prefixed wildcard.

Amongst other things.