Hard-code Fetch-Hosts download dir to %temp%
Previously we allowed the dir parameter, and if the dir did not exist, it was created, removed etc.

This led to constant creation and deletion of the download directory if you were to send host sources individually, instead of part of a collection.

It makes more sense to just output them to the temp dir and delete each one after processing, and fail if the folder is not found.